Refactor DynamicFilter and NavBar components, and update styling
Refactor the DynamicFilter component to fix a bug related to the filterField value. Also, update the NavBar component to add the "place-content-evenly" class for improved layout. Additionally, update the styling in the Pages module to include flex-wrap and a new "card-container-no-wrap" class. These changes enhance the functionality and appearance of the application.